Add complex version of TreeCotreeGaugeEssentialConstraint
