+++ framework/include/mfem/equation_systems/ComplexEquationSystem.h
@@ -9,6 +9,7 @@
 #include "MFEMComplexKernel.h"
 #include "MFEMComplexIntegratedBC.h"
 #include "MFEMComplexEssentialBC.h"
+#include "MFEMComplexEssentialConstraint.h"
 #include "MFEMMixedBilinearFormKernel.h"
 
 namespace Moose::MFEM
@@ -57,6 +58,9 @@ public:
 
   /// Add complex essential BCs
   void AddComplexEssentialBCs(std::shared_ptr<MFEMComplexEssentialBC> bc);
+
+  /// Add complex essential (volumetric) constraints
+  void AddComplexEssentialConstraint(std::shared_ptr<MFEMComplexEssentialConstraint> constraint);
 
   /// Perform trivial eliminations of coupled variables lacking corresponding test variables
   void EliminateCoupledVariables() override;
@@ -124,6 +128,10 @@ protected:
 
   // Complex essential BCs
   NamedFieldsMap<std::vector<std::shared_ptr<MFEMComplexEssentialBC>>> _cmplx_essential_bc_map;
+
+  // Complex essential (volumetric) constraints
+  NamedFieldsMap<std::vector<std::shared_ptr<MFEMComplexEssentialConstraint>>>
+      _cmplx_essential_constraint_map;
 
   /// Pointers to coupled variables not part of the reduced ComplexEquationSystem.
   ComplexGridFunctions _cmplx_eliminated_variables;
