+++ test/tests/mfem/constraints/gold/tree_cotree_gauge_time_harmonic_out.csv
@@ -1,0 +1,3 @@
+time,ConductorField,InsulatorField
+0,0,0
+1,4.5203038706563e-06,0.00013909860653343
